Add in unit tests
That commit splits files by convention one file = one function and add
unit tests to two of them
