+++ package.json
@@ -3,6 +3,9 @@
   "version": "1.1.1",
   "description": "rnpm plugin that links native dependencies to your project",
   "main": "index.js",
+  "scripts": {
+    "test": "mocha ./test --recursive"
+  },
   "keywords": [
     "rnpm",
     "react-native",
@@ -38,6 +41,8 @@
   },
   "devDependencies": {
     "babel-eslint": "^4.1.5",
-    "eslint": "^1.9.0"
+    "chai": "^3.4.1",
+    "eslint": "^1.9.0",
+    "mocha": "^2.3.4"
   }
 }
