+++ src/ios/registerNativeModule.js
@@ -2,15 +2,9 @@ const xcode = require('xcode');
 const fs = require('fs');
 const path = require('path');
 const PbxFile = require('xcode/lib/pbxFile');
-const union = require('lodash.union');
-const glob = require('glob');
-
-const GLOB_EXCLUDE_PATTERN = ['node_modules/**', 'Examples/**', 'examples/**'];
-
-/**
- * Returns last item from an array
- */
-const lastItem = (array) => array[array.length - 1];
+const addToHeaderSearchPaths = require('./addToHeaderSearchPaths');
+const getHeadersInFolder = require('./getHeadersInFolder');
+const getHeaderSearchPath = require('./getHeaderSearchPath');
 
 /**
  * Given an array of libraries already imported and packageName that will be
@@ -56,81 +50,6 @@ const addProjectToLibraries = (libraries, file) => {
     value: file.fileRef,
     comment: file.basename,
   });
-};
-
-/**
- * Given folder, it returns an array of all header files
- * inside it, ignoring node_modules and examples
- */
-const getHeadersInFolder = (folder) =>
-  glob
-    .sync('**/*.h', {
-      cwd: folder,
-      nodir: true,
-      ignore: GLOB_EXCLUDE_PATTERN,
-    })
-    .map(file => path.join(folder, file));
-
-/**
- * Given an array of headers it returns search path so Xcode can find them
- * If there are multiple header files located accross different folders,
- * it finds the top-most one and returns recursive version of it
- */
-const getHeaderSearchPath = (sourceDir, headers) => {
-  const directories = union(
-    headers.map(path.dirname)
-  );
-
-  if (directories.length === 1) {
-    return `"$(SRCROOT)/${path.relative(sourceDir, directories[0])}"`;
-  }
-
-  const directory = directories.reduce((topMostDir, currentDir) => {
-    const currentFolders = currentDir.split(path.sep);
-    const topMostFolders = topMostDir.split(path.sep);
-
-    if (currentFolders.length === topMostFolders.length
-      && lastItem(currentFolders) !== lastItem(topMostFolders)) {
-      return currentFolders.slice(0, -1).join(path.sep);
-    }
-
-    return currentFolders.length < topMostFolders.length
-      ? currentDir
-      : topMostDir;
-  });
-
-  return `"$(SRCROOT)/${path.relative(sourceDir, directory)}/**"`;
-};
-
-
-/**
- * Given Xcode project and path, iterate over all build configurations
- * and append new path to HEADER_SEARCH_PATHS section
- *
- * We cannot use builtin addToHeaderSearchPaths method since react-native init does not
- * use $(TARGET_NAME) for PRODUCT_NAME, but sets it manually so that method will skip
- * that target.
- *
- * To workaround that issue and make it more bullet-proof for different names,
- * we iterate over all configurations and look if React is already there. If it is,
- * we assume user will need newly added dependency headers in it as well.
- */
-const addToHeaderSearchPaths = (project, path) => {
-  const config = project.pbxXCBuildConfigurationSection();
-
-  Object
-    .keys(config)
-    .filter(ref => ref.indexOf('_comment') === -1)
-    .forEach(ref => {
-      const buildSettings = config[ref].buildSettings;
-      const shouldAddSearchPath = (buildSettings.HEADER_SEARCH_PATHS || [])
-        .filter(path => path.indexOf('react-native/React/**'))
-        .length > 0;
-
-      if (shouldAddSearchPath) {
-        buildSettings.HEADER_SEARCH_PATHS.push(path);
-      }
-    });
 };
 
 /**
